Backend: Moving tests into sub-package test-directory
