+++ loki/backend/tests/test_pygen.py
@@ -12,9 +12,11 @@ from collections import namedtuple
 import pytest
 import numpy as np
 
-from loki import Subroutine, FortranPythonTransformation, pygen
+from loki import Subroutine
+from loki.backend import pygen
 from loki.build import jit_compile, clean_test
 from loki.frontend import available_frontends, OFP, OMNI
+from loki.transform import FortranPythonTransformation
 
 
 @pytest.fixture(scope='module', name='here')
